requirements.txt
python-dotenv==1.0.0
pyspark==3.2
psycopg2-binary==2.9.6
matplotlib==3.7.2
pandas==2.0.3
seaborn==0.12.2
